step 2: public catalog types
