+++ crates/store/re_datafusion/src/datafusion_connector.rs
@@ -5,7 +5,7 @@ use tracing::instrument;
 
 use re_grpc_client::ConnectionClient;
 use re_log_types::{EntryId, external::re_tuid::Tuid};
-use re_protos::catalog::v1alpha1::{
+use re_protos::frontend::v1alpha1::{
     DatasetEntry, EntryFilter, ReadDatasetEntryRequest, ext::EntryDetails,
 };
 
@@ -30,7 +30,7 @@ impl DataFusionConnector {
         let entry: EntryDetails = self
             .client
             .inner()
-            .find_entries(re_protos::catalog::v1alpha1::FindEntriesRequest {
+            .find_entries(re_protos::frontend::v1alpha1::FindEntriesRequest {
                 filter: Some(EntryFilter {
                     name: Some("__entries".to_owned()),
                     ..Default::default()
