+++ crates/store/re_grpc_client/src/connection_registry.rs
@@ -6,7 +6,7 @@ use tokio::sync::RwLock;
 use tonic::Code;
 
 use re_auth::Jwt;
-use re_protos::catalog::v1alpha1::{EntryFilter, FindEntriesRequest};
+use re_protos::frontend::v1alpha1::{EntryFilter, FindEntriesRequest};
 
 use crate::TonicStatusError;
 use crate::connection_client::GenericConnectionClient;
